Decimal value fix for POS

+++ src/main/java/com/unisys/trans/cps/middleware/models/response/PointOfSalesResponseDTO.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class PointOfSalesResponseDTO {
 
-    private Long totalValue;
+    private float totalValue;
 
     private float momData;
